2jan

+++ src/main/java/com/scaler/dtos/FakeStoreProductDto.java
@@ -4,12 +4,14 @@ import com.scaler.models.Category;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.UUID;
+
 //class object recieved by ProductService method as response. while using restTemplate object method.
 //jackson library converts json code to object. but we need to give dto object to store response
 @Getter
 @Setter
 public class FakeStoreProductDto {
-    private int id;
+    private Long id;
     private String title;
     private String description;
     private double price;
